demo/requirements.txt
scikit-learn
airflow-provider-flyte
apache-airflow-providers-postgres
